+++ RELEASES.md
@@ -4,10 +4,24 @@ Rhai Release Notes
 Version 0.19.5
 ==============
 
+This version fixes a bug that prevents compilation with the `internals` feature.
+It also speeds up importing modules.
+
+Bug fixes
+---------
+
+* Fixes compilation error when using the `internals` feature.  Bug introduced in `0.19.4`.
+
 Breaking changes
 ----------------
 
 * Modules imported at global level can now be accessed in functions.
+* `ModuleResolver::resolve` now returns `Shared<Module>` for better resources sharing when loading modules.
+
+Enhancements
+------------
+
+* Modules imported via `import` statements at global level can now be used in functions. There is no longer any need to re-`import` the modules at the beginning of each function block.
 
 
 Version 0.19.4